requirements.txt
aiohttp
python-dotenv
py-cord >= 2.0.0
